Extracted common test code
