+++ broker/src/test/java/io/moquette/integration/mqtt5/RequestResponseTest.java
@@ -1,0 +1,2 @@
+package io.moquette.integration.mqtt5;public class RequestResponseTest {
+}
